Move date helper into its own module
We'll be using this in our new script as well.
