+++ dmscripts/notify_buyers_when_requirements_close.py
@@ -6,6 +6,7 @@ from dmutils.email import send_email
 from dmutils.email.exceptions import EmailError
 from dmutils.formats import DATE_FORMAT, DATETIME_FORMAT
 
+from dmscripts.helpers.date_helpers import get_date_closed
 from dmscripts.helpers.email_helpers import get_sent_emails
 from dmscripts.helpers.html_helpers import render_html
 from dmscripts.helpers import logging_helpers
@@ -52,13 +53,6 @@ def notify_users(email_api_key, brief):
             )
 
             return False
-
-
-def get_date_closed(date_closed):
-    if date_closed is None:
-        return (datetime.utcnow() - timedelta(days=1)).date()
-    else:
-        return datetime.strptime(date_closed, DATE_FORMAT).date()
 
 
 def get_notified_briefs(email_api_key, date_closed):
