+++ tests/test_notify_buyers_when_requirements_close.py
@@ -5,7 +5,6 @@ import datetime
 
 from dmscripts.notify_buyers_when_requirements_close import (
     get_closed_briefs,
-    get_date_closed,
     get_notified_briefs,
     notify_users,
     main
@@ -48,18 +47,6 @@ def test_get_closed_briefs_filters_by_date_closed():
         {"applicationsClosedAt": "2016-09-05T23:59:59.000000Z", "status": "closed"},
         {"applicationsClosedAt": "2016-09-05T23:59:59.000000Z", "status": "closed"},
     ]
-
-
-def test_get_date_closed():
-    def check_date_closed(value, expected):
-        with freeze_time('2015-01-02 03:04:05'):
-            assert get_date_closed(value) == expected
-
-    for value, expected in [
-        (None, datetime.date(2015, 1, 1)),
-        ('2016-01-02', datetime.date(2016, 1, 2))
-    ]:
-        yield check_date_closed, value, expected
 
 
 @mock.patch('dmscripts.notify_buyers_when_requirements_close.get_sent_emails')
